extrayendo el calculo del proximo dia de root a una clase
